fix(moe): re-apply GPTQ Marlin MoE bf16 scale dtype support (fork 232ce582f)

+++ python/sglang/srt/layers/quantization/gptq/schemes/gptq_moe.py
@@ -205,7 +205,7 @@ class GPTQMarlinMoEScheme(GPTQMoESchemeBase):
                 num_experts,
                 scales_size13,
                 2 * intermediate_size_per_partition,
-                dtype=torch.half,
+                dtype=params_dtype,
             ),
             requires_grad=False,
         )
@@ -213,7 +213,7 @@ class GPTQMarlinMoEScheme(GPTQMoESchemeBase):
         set_weight_attrs(w13_scales, extra_weight_attrs)
 
         w2_scales = torch.nn.Parameter(
-            torch.empty(num_experts, scales_size2, hidden_size, dtype=torch.half),
+            torch.empty(num_experts, scales_size2, hidden_size, dtype=params_dtype),
             requires_grad=False,
         )
         layer.register_parameter("w2_scales", w2_scales)
